taskmanager/requirements.txt
asgiref==3.11.0
cffi==2.0.0
cryptography==46.0.3
Django==6.0.1
django-environ==0.13.0
django-taggit==6.1.0
djangorestframework==3.16.1
gunicorn==25.1.0
mysql-connector-python==9.5.0
mysqlclient==2.2.7
packaging==26.0
pillow==12.1.0
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
pycparser==2.23
sqlparse==0.5.5
tzdata==2025.3
whitenoise==6.11.0
